finish elective 1
