+++ lab2/bootloader/linker.ld
@@ -1,10 +1,11 @@
-ENTRY(_start)
+ENTRY(_boot)
 
-__stack_top = 0x92000;
+__stack_top  = 0x90000;
 
 SECTIONS
 {
     . = 0x80000;
+    __bootlaoder_start = .;
     .text : { KEEP(*(.text.boot)) }
     .text :  { *(.text) }
     .rodata : { *(.rodata) }
@@ -15,3 +16,4 @@ SECTIONS
     __bss_end = .;
 }
 __bss_size = (__bss_end - __bss_start) >> 3;
+__bootloader_size = (__bss_end - __bootlaoder_start) >> 3;
