+++ lab2/config.txt
@@ -1,4 +1,3 @@
-kernel_address=0x60000
 kernel=bootloader.img
 arm64_bit=1
 initramfs initramfs.cpio 0x20000000
