+++ lab2/send.py
@@ -1,5 +1,6 @@
 import os
 import argparse
+import time
 
 parser = argparse.ArgumentParser()
 parser.add_argument('--path', '-p', type=str, default='kernel8.img')
@@ -8,6 +9,7 @@ args = parser.parse_args()
 
 path = "build/{}".format(args.path)
 PORT = "/dev/pts/{}".format(args.port)
+#PORT = "/dev/ttyUSB0"
 
 size = os.path.getsize(path)
 print((str(size)).encode())
@@ -16,6 +18,7 @@ with open(PORT, "wb", buffering=0) as tty:
     #print(str(file_size)+'\n')
     tty.write((str(size)+'\0').encode())
 
+#flag = 0
 with open(path, "rb", buffering=0) as f:
    while 1:
         with open(PORT, "wb", buffering=0) as tty:
@@ -23,5 +26,8 @@ with open(path, "rb", buffering=0) as f:
             if not byte:
                 break
             tty.write(byte)
+        #print(flag)
+        #flag += 1
+        #time.sleep(0.001)
 
 print("Done")
