--- lab2/src/boot.S
@@ -1,38 +1,0 @@
-.section ".text.boot"
-
-.global _boot
-
-_boot:
-    // read cpu id, stop slave cores
-    mrs x1, mpidr_el1
-    and x1, x1, #3
-    cbz x1, 2f
-    // cpu id > 0, stop
-1:  
-    wfe
-    b 1b
-2:  
-    // set stack before our code
-    ldr x1, =__stack_top
-    mov sp, x1
-
-    // clear bss
-    ldr x1, =__bss_start
-    ldr x2, =__bss_size
-
-clear_bss:
-    cbz x2, 4f
-
-    // post-indexed store
-    str xzr, [x1], #8
-    sub x2, x2, #1
-    cbnz x2, clear_bss
-
-4:  
-    bl secondBootloader
-
-//5:
-//    b 5b //branch to 5 back
-
-
-
