--- lab2/src/bootloader.h
@@ -1,1 +1,0 @@
-void secondBootloader();
